refactor: 残り4ファイルの責務分離（第2弾）
calendar-group-manager.js (715→465行): グループモーダルのDOM生成を
GroupModalBuilderに切り出し。

google-event-modal.js (694→369行): イベント詳細の各セクション生成を
GoogleEventContentBuilderに切り出し。

memo-component.js (666→329行): マークダウン編集ロジック（キーボード操作、
レンダリング、チェックボックス切替）をMemoEditorに切り出し。

timeline-calendar-filter.js (797→約430行): ドロップダウン描画ロジックを
CalendarFilterRendererに切り出し。

全て純粋なリファクタ（機能変更なし）。501テスト全通過。
