+++ src/side_panel/components/timeline/timeline-calendar-filter.js
@@ -7,6 +7,7 @@ import { Component } from '../base/component.js';
 import { sendMessage } from '../../../lib/chrome-messaging.js';
 import { loadSelectedCalendars, saveSelectedCalendars, loadCalendarGroups } from '../../../lib/settings-storage.js';
 import { isDemoMode, getDemoCalendarGroups } from '../../../lib/demo-data.js';
+import { CalendarFilterRenderer } from './calendar-filter-renderer.js';
 
 export class TimelineCalendarFilter extends Component {
     constructor(options = {}) {
@@ -36,6 +37,22 @@ export class TimelineCalendarFilter extends Component {
         this._boundOnScroll = null;
         this._rafId = null;
         this._isFetching = false;
+
+        // Renderer delegate
+        this.renderer = new CalendarFilterRenderer({
+            getCalendars: () => this.calendars,
+            getSelectedIds: () => this.selectedIds,
+            getCalendarGroups: () => this.calendarGroups,
+            getSearchTerm: () => this.searchTerm,
+            getMessage: (key) => this.getMessage(key),
+            onSearchInput: (value) => {
+                this.searchTerm = value;
+                this.renderer.renderCalendarList();
+            },
+            onRefreshClick: () => this._refreshCalendars(),
+            onCalendarToggle: (calendarId, checked) => this._handleToggle(calendarId, checked),
+            onGroupToggle: (group, calendars, checked) => this._handleGroupToggle(group, calendars, checked),
+        });
     }
 
     createElement() {
@@ -315,308 +332,14 @@ export class TimelineCalendarFilter extends Component {
     }
 
     /**
-     * Render the full dropdown content (toolbar + calendar list)
+     * Delegate rendering to CalendarFilterRenderer, then sync DOM refs back.
      * @private
      */
     _renderDropdownContent() {
-        this.dropdown.innerHTML = '';
-
-        // Toolbar: search + refresh
-        const toolbar = document.createElement('div');
-        toolbar.className = 'timeline-calendar-filter-toolbar';
-
-        this.searchInput = document.createElement('input');
-        this.searchInput.type = 'text';
-        this.searchInput.className = 'timeline-calendar-filter-search';
-        this.searchInput.placeholder = this.getMessage('calendarFilterSearchPlaceholder');
-        this.searchInput.setAttribute('aria-label', this.getMessage('calendarFilterSearchPlaceholder'));
-        this.searchInput.value = this.searchTerm;
-        this.searchInput.addEventListener('input', () => {
-            this.searchTerm = this.searchInput.value;
-            this._renderCalendarList();
-        });
-
-        this.refreshBtn = document.createElement('button');
-        this.refreshBtn.type = 'button';
-        this.refreshBtn.className = 'timeline-calendar-filter-refresh-btn';
-        this.refreshBtn.title = this.getMessage('calendarFilterRefreshTooltip');
-        this.refreshBtn.setAttribute('aria-label', this.getMessage('calendarFilterRefreshTooltip'));
-        this.refreshBtn.innerHTML = '<i class="fa-solid fa-arrows-rotate"></i>';
-        this.refreshBtn.addEventListener('click', () => {
-            this._refreshCalendars();
-        });
-
-        toolbar.appendChild(this.searchInput);
-        toolbar.appendChild(this.refreshBtn);
-        this.dropdown.appendChild(toolbar);
-
-        // Calendar list container
-        this.calendarList = document.createElement('div');
-        this.calendarList.className = 'timeline-calendar-filter-list';
-        this.dropdown.appendChild(this.calendarList);
-
-        this._renderCalendarList();
-    }
-
-    /**
-     * Render the calendar list with group support
-     * @private
-     */
-    _renderCalendarList() {
-        if (!this.calendarList) return;
-        this.calendarList.innerHTML = '';
-
-        if (!this.calendars || this.calendars.length === 0) {
-            const empty = document.createElement('div');
-            empty.className = 'timeline-calendar-filter-status';
-            empty.textContent = this.getMessage('noCalendarsAvailable');
-            this.calendarList.appendChild(empty);
-            return;
-        }
-
-        // Filter by search term
-        const term = this.searchTerm.toLowerCase().trim();
-        const filtered = term
-            ? this.calendars.filter(c => (c.summary || '').toLowerCase().includes(term))
-            : this.calendars;
-
-        if (filtered.length === 0) {
-            const noResult = document.createElement('div');
-            noResult.className = 'timeline-calendar-filter-status';
-            noResult.textContent = this.getMessage('noCalendarsAvailable');
-            this.calendarList.appendChild(noResult);
-            return;
-        }
-
-        const filteredIds = new Set(filtered.map(c => c.id));
-        const calendarMap = new Map(this.calendars.map(c => [c.id, c]));
-
-        // If groups exist, render grouped
-        if (this.calendarGroups.length > 0) {
-            for (const group of this.calendarGroups) {
-                const groupCalendars = group.calendarIds
-                    .map(id => calendarMap.get(id))
-                    .filter(cal => cal && filteredIds.has(cal.id));
-
-                if (term && groupCalendars.length === 0) continue;
-
-                this._renderGroupSection(group, groupCalendars);
-            }
-
-            // Ungrouped
-            const groupedIds = new Set();
-            for (const group of this.calendarGroups) {
-                for (const id of group.calendarIds) {
-                    groupedIds.add(id);
-                }
-            }
-
-            const ungrouped = filtered
-                .filter(c => !groupedIds.has(c.id))
-                .sort((a, b) => {
-                    if (a.primary && !b.primary) return -1;
-                    if (!a.primary && b.primary) return 1;
-                    return (a.summary || '').localeCompare(b.summary || '');
-                });
-
-            if (ungrouped.length > 0) {
-                this._renderUngroupedSection(ungrouped);
-            }
-        } else {
-            // No groups - flat list (backward compatible)
-            const sorted = [...filtered].sort((a, b) => {
-                if (a.primary && !b.primary) return -1;
-                if (!a.primary && b.primary) return 1;
-                return (a.summary || '').localeCompare(b.summary || '');
-            });
-
-            sorted.forEach(calendar => {
-                this._renderCalendarItem(this.calendarList, calendar);
-            });
-        }
-    }
-
-    /**
-     * Render a group section in the filter dropdown
-     * @private
-     */
-    _renderGroupSection(group, calendars) {
-        // Group header
-        const header = document.createElement('div');
-        header.className = 'timeline-calendar-filter-group-header';
-        header.dataset.groupId = group.id;
-
-        header.setAttribute('role', 'button');
-        header.setAttribute('tabindex', '0');
-        header.setAttribute('aria-expanded', 'true');
-
-        const checkbox = document.createElement('input');
-        checkbox.type = 'checkbox';
-        checkbox.setAttribute('aria-label', group.name);
-
-        // Determine check state using full group membership (not search-filtered view)
-        const calendarMap = new Map(this.calendars.map(c => [c.id, c]));
-        const fullGroupIds = group.calendarIds.filter(id => calendarMap.has(id));
-        const selectedCount = fullGroupIds.filter(id => this.selectedIds.includes(id)).length;
-        if (selectedCount === 0 || fullGroupIds.length === 0) {
-            checkbox.checked = false;
-            checkbox.setAttribute('aria-checked', 'false');
-        } else if (selectedCount === fullGroupIds.length) {
-            checkbox.checked = true;
-            checkbox.setAttribute('aria-checked', 'true');
-        } else {
-            checkbox.checked = false;
-            checkbox.indeterminate = true;
-            checkbox.setAttribute('aria-checked', 'mixed');
-        }
-
-        checkbox.addEventListener('change', (e) => {
-            e.stopPropagation();
-            this._handleGroupToggle(group, calendars, checkbox.checked);
-        });
-
-        const nameSpan = document.createElement('span');
-        nameSpan.className = 'group-name';
-        nameSpan.textContent = group.name;
-
-        const collapseIcon = document.createElement('i');
-        collapseIcon.className = 'fa-solid fa-chevron-down group-collapse-icon';
-        collapseIcon.setAttribute('aria-hidden', 'true');
-
-        header.appendChild(checkbox);
-        header.appendChild(nameSpan);
-        header.appendChild(collapseIcon);
-
-        const toggleCollapse = () => {
-            const body = header.nextElementSibling;
-            if (body) body.classList.toggle('collapsed');
-            collapseIcon.classList.toggle('collapsed');
-            const expanded = !collapseIcon.classList.contains('collapsed');
-            header.setAttribute('aria-expanded', String(expanded));
-        };
-
-        header.addEventListener('click', (e) => {
-            if (e.target === checkbox) return;
-            toggleCollapse();
-        });
-
-        header.addEventListener('keydown', (e) => {
-            if (e.key === 'Enter' || e.key === ' ') {
-                if (e.target === checkbox) return;
-                e.preventDefault();
-                toggleCollapse();
-            }
-        });
-
-        this.calendarList.appendChild(header);
-
-        // Group body
-        const body = document.createElement('div');
-        body.className = 'timeline-calendar-filter-group-body';
-
-        const sorted = [...calendars].sort((a, b) => {
-            if (a.primary && !b.primary) return -1;
-            if (!a.primary && b.primary) return 1;
-            return (a.summary || '').localeCompare(b.summary || '');
-        });
-
-        sorted.forEach(calendar => {
-            this._renderCalendarItem(body, calendar);
-        });
-
-        this.calendarList.appendChild(body);
-    }
-
-    /**
-     * Render ungrouped section
-     * @private
-     */
-    _renderUngroupedSection(calendars) {
-        const header = document.createElement('div');
-        header.className = 'timeline-calendar-filter-group-header';
-        header.setAttribute('role', 'button');
-        header.setAttribute('tabindex', '0');
-        header.setAttribute('aria-expanded', 'true');
-
-        const nameSpan = document.createElement('span');
-        nameSpan.className = 'group-name';
-        nameSpan.textContent = this.getMessage('ungrouped') || 'Ungrouped';
-
-        const collapseIcon = document.createElement('i');
-        collapseIcon.className = 'fa-solid fa-chevron-down group-collapse-icon';
-        collapseIcon.setAttribute('aria-hidden', 'true');
-
-        header.appendChild(nameSpan);
-        header.appendChild(collapseIcon);
-
-        const toggleCollapse = () => {
-            const body = header.nextElementSibling;
-            if (body) body.classList.toggle('collapsed');
-            collapseIcon.classList.toggle('collapsed');
-            const expanded = !collapseIcon.classList.contains('collapsed');
-            header.setAttribute('aria-expanded', String(expanded));
-        };
-
-        header.addEventListener('click', () => toggleCollapse());
-
-        header.addEventListener('keydown', (e) => {
-            if (e.key === 'Enter' || e.key === ' ') {
-                e.preventDefault();
-                toggleCollapse();
-            }
-        });
-
-        this.calendarList.appendChild(header);
-
-        const body = document.createElement('div');
-        body.className = 'timeline-calendar-filter-group-body';
-
-        calendars.forEach(calendar => {
-            this._renderCalendarItem(body, calendar);
-        });
-
-        this.calendarList.appendChild(body);
-    }
-
-    /**
-     * Render a single calendar item
-     * @private
-     */
-    _renderCalendarItem(container, calendar) {
-        const isSelected = this.selectedIds.includes(calendar.id);
-        const item = document.createElement('label');
-        item.className = 'timeline-calendar-filter-item';
-
-        // Checkbox
-        const checkbox = document.createElement('input');
-        checkbox.type = 'checkbox';
-        checkbox.className = 'timeline-calendar-filter-checkbox';
-        checkbox.checked = isSelected || calendar.primary;
-        if (calendar.primary) {
-            checkbox.disabled = true;
-        }
-        checkbox.addEventListener('change', () => {
-            this._handleToggle(calendar.id, checkbox.checked);
-        });
-
-        // Color indicator
-        const color = document.createElement('span');
-        color.className = 'timeline-calendar-filter-color';
-        color.style.backgroundColor = calendar.backgroundColor || '#ccc';
-        color.setAttribute('aria-hidden', 'true');
-
-        // Name
-        const name = document.createElement('span');
-        name.className = 'timeline-calendar-filter-name';
-        name.textContent = calendar.summary;
-        if (calendar.primary) {
-            name.classList.add('timeline-calendar-filter-primary');
-        }
-
-        item.appendChild(checkbox);
-        item.appendChild(color);
-        item.appendChild(name);
-        container.appendChild(item);
+        const refs = this.renderer.renderDropdownContent(this.dropdown);
+        this.searchInput = refs.searchInput;
+        this.refreshBtn = refs.refreshBtn;
+        this.calendarList = refs.calendarList;
     }
 
     /**
@@ -660,7 +383,7 @@ export class TimelineCalendarFilter extends Component {
 
         try {
             await saveSelectedCalendars(this.selectedIds);
-            this._renderCalendarList();
+            this.renderer.renderCalendarList();
             if (this.onCalendarChange) {
                 const addedIds = this.selectedIds.filter(id => !previousIds.includes(id));
                 const removedIds = previousIds.filter(id => !this.selectedIds.includes(id));
@@ -668,7 +391,7 @@ export class TimelineCalendarFilter extends Component {
             }
         } catch {
             this.selectedIds = previousIds;
-            this._renderCalendarList();
+            this.renderer.renderCalendarList();
         }
     }
 
@@ -712,7 +435,7 @@ export class TimelineCalendarFilter extends Component {
         try {
             await saveSelectedCalendars(this.selectedIds);
             // Update group header checkbox states
-            this._updateGroupCheckboxStates();
+            this.renderer.updateGroupCheckboxStates(this.calendarList);
             if (this.onCalendarChange) {
                 this.onCalendarChange({
                     addedIds: checked ? [calendarId] : [],
@@ -721,41 +444,7 @@ export class TimelineCalendarFilter extends Component {
             }
         } catch {
             this.selectedIds = previousIds;
-            this._renderCalendarList();
-        }
-    }
-
-    /**
-     * Update group header checkbox states after individual toggle
-     * @private
-     */
-    _updateGroupCheckboxStates() {
-        if (!this.calendarList || this.calendarGroups.length === 0) return;
-
-        const calendarMap = new Map(this.calendars.map(c => [c.id, c]));
-
-        for (const group of this.calendarGroups) {
-            const header = this.calendarList.querySelector(`.timeline-calendar-filter-group-header[data-group-id="${CSS.escape(group.id)}"]`);
-            if (!header) continue;
-
-            const checkbox = header.querySelector('input[type="checkbox"]');
-            if (!checkbox) continue;
-
-            const validCalendars = group.calendarIds.filter(id => calendarMap.has(id));
-            const selectedCount = validCalendars.filter(id => this.selectedIds.includes(id)).length;
-
-            checkbox.indeterminate = false;
-            if (selectedCount === 0) {
-                checkbox.checked = false;
-                checkbox.setAttribute('aria-checked', 'false');
-            } else if (selectedCount === validCalendars.length) {
-                checkbox.checked = true;
-                checkbox.setAttribute('aria-checked', 'true');
-            } else {
-                checkbox.checked = false;
-                checkbox.indeterminate = true;
-                checkbox.setAttribute('aria-checked', 'mixed');
-            }
+            this.renderer.renderCalendarList();
         }
     }
 
@@ -792,6 +481,7 @@ export class TimelineCalendarFilter extends Component {
         this.calendars = [];
         this.selectedIds = [];
         this.calendarGroups = [];
+        this.renderer = null;
         super.destroy();
     }
 }
